Make the struct compatible with the response from the server

+++ src/methods/send_chat_action.rs
@@ -3,7 +3,7 @@ use types::ChatAction;
 
 #[derive(Debug, Builder, Serialize)]
 pub struct SendChatAction {
-    chat_id: i32,
+    chat_id: i64,
     action: ChatAction,
 }
